Alternative approach to cookie helpers
An attempt to avoid the need for CurrentAttributes by pushing the
controller helper logic into the concern. Used a consistent helpers
concern to give us the same interface as the design system equivalent
but this is just a style preference. Combines the specs so we're setting
the cookie value as part of the test.
