--- engine/app/controllers/concerns/citizens_advice_cookie_preferences/set_cookie_preferences.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-module CitizensAdviceCookiePreferences
-  module SetCookiePreferences
-    extend ActiveSupport::Concern
-
-    included do
-      before_action :set_cookie_preferences
-    end
-
-    protected
-
-    def set_cookie_preferences
-      return if cookies[:cookie_preference].blank?
-
-      CurrentCookies.preference = JSON.parse(cookies[:cookie_preference])
-    end
-  end
-end
