+++ engine/app/helpers/citizens_advice_cookie_preferences/application_helper.rb
@@ -1,13 +1,5 @@
 # frozen_string_literal: true
 
 module CitizensAdviceCookiePreferences
-  module ApplicationHelper
-    def allow_analytics_cookies?
-      CitizensAdviceCookiePreferences::CurrentCookies.analytics?
-    end
-
-    def allow_video_players_cookies?
-      CitizensAdviceCookiePreferences::CurrentCookies.video_players?
-    end
-  end
+  module ApplicationHelper; end
 end
