--- engine/app/models/citizens_advice_cookie_preferences/current_cookies.rb
@@ -1,23 +1,0 @@
-# frozen_string_literal: true
-
-module CitizensAdviceCookiePreferences
-  class CurrentCookies < ActiveSupport::CurrentAttributes
-    attribute :preference
-
-    def analytics?
-      if preference.present?
-        preference.fetch("analytics", false)
-      else
-        false
-      end
-    end
-
-    def video_players?
-      if preference.present?
-        preference.fetch("video_players", false)
-      else
-        false
-      end
-    end
-  end
-end
