+++ engine/spec/controllers/citizens_advice_cookie_preferences/application_controller_spec.rb
@@ -2,25 +2,32 @@
 
 require "rails_helper"
 
-RSpec.describe CitizensAdviceCookiePreferences::CurrentCookies do
-  subject(:current_cookies) { described_class.new }
+RSpec.describe ApplicationController do
+  controller do
+    include CitizensAdviceCookiePreferences::Helpers
+
+    def index
+      render plain: "I like cookies!"
+    end
+  end
+
+  before do
+    request.cookies[:cookie_preference] = preference.to_json
+    get :index
+  end
 
   let(:preference) { nil }
 
-  before do
-    current_cookies.preference = preference
-  end
-
   describe "#analytics?" do
     it "returns false when no cookies are set" do
-      expect(current_cookies.analytics?).to be false
+      expect(controller.helpers.allow_analytics_cookies?).to be false
     end
 
     context "with analytics cookies accepted" do
       let(:preference) { { "essential" => true, "analytics" => true, "video_players" => false } }
 
       it "returns true when the analytics cookie is true" do
-        expect(current_cookies.analytics?).to be true
+        expect(controller.helpers.allow_analytics_cookies?).to be true
       end
     end
 
@@ -28,21 +35,21 @@ RSpec.describe CitizensAdviceCookiePreferences::CurrentCookies do
       let(:preference) { { "essential" => true, "analytics" => false, "video_players" => false } }
 
       it "returns false when the analytics cookie is false" do
-        expect(current_cookies.analytics?).to be false
+        expect(controller.helpers.allow_analytics_cookies?).to be false
       end
     end
   end
 
   describe "#video_players?" do
     it "returns false when no cookies are set" do
-      expect(current_cookies.video_players?).to be false
+      expect(controller.helpers.allow_video_players_cookies?).to be false
     end
 
     context "with video_players cookies accepted" do
       let(:preference) { { "essential" => true, "analytics" => false, "video_players" => true } }
 
       it "returns true when the video_players cookie is true" do
-        expect(current_cookies.video_players?).to be true
+        expect(controller.helpers.allow_video_players_cookies?).to be true
       end
     end
 
@@ -50,7 +57,7 @@ RSpec.describe CitizensAdviceCookiePreferences::CurrentCookies do
       let(:preference) { { "essential" => true, "analytics" => false, "video_players" => false } }
 
       it "returns false when the video_players cookie is false" do
-        expect(current_cookies.video_players?).to be false
+        expect(controller.helpers.allow_video_players_cookies?).to be false
       end
     end
   end
